// packages/shared/src/jonasland/deployment/fly-api/generated/openapi.gen.ts
export interface paths {
  "/apps": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * List Apps
     * @description List all apps with the ability to filter by organization slug.
     */
    get: operations["Apps_list"];
    put?: never;
    /**
     * Create App
     * @description Create an app with the specified details in the request body.
     */
    post: operations["Apps_create"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/apps/{app_name}": {
    parameters: {
      query?: never;
      header?: never;
      path: {
        app_name: string;
      };
      cookie?: never;
    };
    /**
     * Get App
     * @description Retrieve details about a specific app by its name.
     */
    get: operations["Apps_show"];
    put?: never;
    post?: never;
    /**
     * Destroy App
     * @description Delete an app by its name.
     */
    delete: operations["Apps_delete"];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/apps/{app_name}/ip_assignments": {
    parameters: {
      query?: never;
      header?: never;
      path: {
        app_name: string;
      };
      cookie?: never;
    };
    /** List IP assignments for app */
    get: operations["App_IPAssignments_list"];
    put?: never;
    /** Assign new IP address to app */
    post: operations["App_IPAssignments_create"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/apps/{app_name}/machines": {
    parameters: {
      query?: never;
      header?: never;
      path: {
        app_name: string;
      };
      cookie?: never;
    };
    /**
     * List Machines
     * @description List all Machines associated with a specific app, with optional filters for including deleted Machines and filtering by region.
     */
    get: operations["Machines_list"];
    put?: never;
    /**
     * Create Machine
     * @description Create a Machine within a specific app using the details provided in the request body.
     *
     *     **Important**: This request can fail, and you’re responsible for handling that failure. If you ask for a large Machine, or a Machine in a region we happen to be at capacity for, you might need to retry the request, or to fall back to another region. If you’re working directly with the Machines API, you’re taking some responsibility for your own orchestration!
     */
    post: operations["Machines_create"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  "/apps/{app_name}/machines/{machine_id}": {
    parameters: {
      query?: never;
      header?: never;
      path: {
        app_name: string;
        machine_id: string;
      };
      cookie?: never;
    };
    /**
     * Get Machine
     * @description Get details of a specific Machine within an app by the Machine ID.
     */
    get: operations["Machines_show"];
    put?: never;
    /**
     * Update Machine
     * @description Update a Machine's configuration using the details provided in the request body.
     */
    post: operations["Machines_update"];
    /**
     * Destroy Machine
     * @description Delete a specific Machine within an app by Machine ID, with an optional force parameter to force kill the Machine if it's running.
     */
    delete: operations["Machines_delete"];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/apps/{app_name}/machines/{machine_id}/exec": {
    parameters: {
      query?: never;
      header?: never;
      path: {
        app_name: string;
        machine_id: string;
      };
      cookie?: never;
    };
    get?: never;
    put?: never;
    /**
     * Execute Command
     * @description Execute a command on a specific Machine and return the raw command output bytes.
     */
    post: operations["Machines_exec"];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  "/apps/{app_name}/machines/{machine_id}/wait": {
    parameters: {
      query?: never;
      header?: never;
      path: {
        app_name: string;
        machine_id: string;
      };
      cookie?: never;
    };
    /**
     * Wait for State
     * @description Wait for a Machine to reach a specific state. Specify the desired state with the state parameter. See the [Machine states table](https://fly.io/docs/machines/working-with-machines/#machine-states) for a list of possible states. The default for this parameter is `started`.
     *
     *     This request will block for up to 60 seconds. Set a shorter timeout with the timeout parameter.
     */
    get: operations["Machines_wait"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
// ...
    CreateMachineRequest: {
      /** @description An object defining the Machine configuration */
      config?: components["schemas"]["fly.MachineConfig"];
      lease_ttl?: number;
      lsvd?: boolean;
      min_secrets_version?: number;
      /** @description Unique name for this Machine. If omitted, one is generated for you */
      name?: string;
      /** @description The target region. Omitting this param launches in the same region as your WireGuard peer connection (somewhere near you). */
      region?: string;
      skip_launch?: boolean;
      skip_secrets?: boolean;
      skip_service_registration?: boolean;
    };
// ...
    Machine: {
// ...
      state?: string;
      updated_at?: string;
    };
// ...
    "fly.ContainerHealthcheck": {
      exec?: components["schemas"]["fly.ExecHealthcheck"];
      /** @description The number of times the check must fail before considering the container unhealthy. */
      failure_threshold?: number;
      /** @description The time in seconds to wait after a container starts before checking its health. */
      grace_period?: number;
      http?: components["schemas"]["fly.HTTPHealthcheck"];
// ...
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}